Rename Vanilla to 'Flexi'.
