+++ quizz_question/quizz_question.hooks.inc
@@ -177,11 +177,11 @@ function quizz_question_quizz_question_info() {
       'module'            => 'quizz_question',
   );
 
-  $info['vanilla'] = array(
-      'name'                  => t('Vanilla'),
+  $info['flexi'] = array(
+      'name'                  => t('Flexi'),
       'description'           => t('Question type with no fields by default, do it yourself.'),
-      'question provider'     => 'Drupal\quizz_question\Handler\Vanilla\VanillaQuestionHandler',
-      'response provider'     => 'Drupal\quizz_question\Handler\Vanilla\VanillaResonseHandler',
+      'question provider'     => 'Drupal\quizz_question\Handler\Flexi\FlexiQuestionHandler',
+      'response provider'     => 'Drupal\quizz_question\Handler\Flexi\FlexiResonseHandler',
       'create default bundle' => FALSE,
       'module'                => 'quizz_question',
   );
